modularizing homepage - qoa
